+++ src/app/core/data/terminalData.js
