+++ src/app/landing/About.jsx
@@ -3,88 +3,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faEnvelope } from '@fortawesome/free-solid-svg-icons'
 import { faInstagram, faLinkedinIn } from '@fortawesome/free-brands-svg-icons'
 import img from '../../assets/img/me.png'
-
-const INFO = [
-  "My name is Moises Arrona I'm Software Engineer and Software Architect and I worked with any lenguages I know about BackEnd and FrontEnd, I currently live in Mexico and I work as an Software Architect",
-  "I like to learn about technology, programming lenguages, software architects, technology tools and I create projects open source as a hobby, well I also like to play guitar.",
-  "I worked at Werkn as a Frontend Developer, Grupo Castores as a Software Engineer, currently working at Flecha Amarilla as a Software Architect. My activities at GFA are: proposing technological solutions, architectural solutions to migrate systems and for new developments. Create and Propose guidelines, design patterns, standards. Provide research and technical training. Develop proofs of concept."
-]
-
-const SKILLS = [
-  {
-    name: "Java / SpringBoot",
-    percennt: "90%"
-  },
-  {
-    name: "C# / .ASP.NET",
-    percennt: "50%"
-  },
-  {
-    name: "PHP / Laravel",
-    percennt: "95%"
-  },
-  {
-    name: "React",
-    percennt: "60%"
-  },
-  {
-    name: "Angular",
-    percennt: "90%"
-  },
-  {
-    name: "Git / GitHub",
-    percennt: "90%"
-  },
-  {
-    name: "SQLServer / MySql",
-    percennt: "80%"
-  }
-]
-
-const STUDIES = [
-  {
-    id: 1,
-    academy: "Universidad Tecnológica de León",
-    // period: "Sept 2016 - Ago 2018",
-    period: "2016 - 2018",
-    title: "TSU en Multimedia y Comercio Electrónico"
-  },
-  {
-    id: 2,
-    academy: "Universidad Tecnológica de León",
-    // period: "Sept 2018 - Abr 2020",
-    period: "2018 - 2020",
-    title: "Ingeniería en Tecnologías de la Información"
-  },
-]
-
-const CONTACT = [
-  {
-    id: 1,
-    name: "instagram",
-    link: "https://www.instagram.com/moisesarrona/",
-    active: true
-  },
-  {
-    id: 1,
-    name: "linkedin",
-    link: "https://www.linkedin.com/in/moisesarrona/",
-    active: true
-  },
-  {
-    id: 1,
-    name: "github",
-    link: "https://github.com/moisesarrona",
-    icon: "",
-    active: true
-  },
-  {
-    id: 1,
-    name: "correo",
-    link: "arronamoisesar@gmail.com",
-    active: true
-  }
-]
+import { INFO, SKILLS, STUDIES, CONTACT } from '../core/data/aboutData'
 
 const About = () => {
   return (
